requirements.txt
fastapi
pydantic-settings
uvicorn[standard]
sqlalchemy
databases[aiosqlite]
psycopg2-binary
psycopg2
#databases[asycpg]
python-dotenv
pydantic-settings
rich
asgi-correlation-id
python-json-logger
python-multipart
passlib[bcrypt]
python-jose[cryptography]
httpx
aiofiles
b2sdk